qbittorrent: geoip fix
Turns out the link on the geoip website moves around every couple of
weeks, and that was breaking people’s installs using this script. I’ve
switched to using the Debian versioned downloads for now, as they won’t
move around. Not a perfect solution, and it’s one I’d like to move away
from eventually, but it works for now.

Also added some changes to the stable build to make compile almost
work. Still not quite there, but progress.