truecrypt: removal
For now, Truecrypt is getting removed. The tests are failing during
compile and during runtime, which makes creating new encrypted
containers impossible. I'm working on narrowing down the particular
test that is failing, but until I find it and potentially fix it,
Truecrypt is going to remain broken.

If your need is urgent, I recommend http://sourceforge.net/projects/veracrypt/
as a stopgap.

See #2 for background.